hypervisor: Create an exit status serial device for the executed process
We want to exfiltrate the process exit status out of the VM to give it
back to the user.

A named pipe to transport the exit status is created by containerd-shim
and we hook a virto serial device onto it. The "create" command has a
new option to tell us where to find that pipe.

Then, inside the VM, we write the workload exit status into that virtio
serial device. containerd-shim will read that value at the other end of
the pipe and forward it to containerd.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>