proxy: Add a `bye` command
We need a way to detect that the client has finished with the
connection, either by detecting the peer has closed the socket or by a
more active way (eg. a bye command or keep alive packets).

It's not that easy to detect that the peer has closed the socket when
using the high level Go net functions. I'd use epoll() and EPOLLRDHUP
but there isn't really a way to plug it in.

So instead, let's try the `bye` command approach, at least for now.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>