proxy: Use a struct to encapsulate the return state of a Handler
At the moment protocol handlers can return an error and a map that will
serialized to JSON in the answer.

I now would like to add the possibility of returning a file descriptor
as part of a protocol Response.

For that to be cleanly done, what the handler can return needs to be a
bit more than just a list of return values and it's time to introduce a
struct to store the states handlers can return.

This changes the handler API from:

func echoHandler(data []byte, userData interface{}) (map[string]interface{}, error) {
	/* do things *.
       return map[string]interface{}{
               "key": value,
       }, nil
}

To:

func echoHandler(data []byte, userData interface{}, response *HandlerResponse) {
	/* do things *.
	response.AddResult("result", echo.Arg)
}

This way, we can keep adding more things to the response struct without being
too invasive.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>