tests: Run the functional tests from builddir in out of tree builds
We have slight issue when srcdir != builddi (out of tree builds)
because, then, common.bash ends up in builddir (generated file) while
the .bats files are in srcdir.

Running bats with the above, the "load common" directive fails because
bats expects common.bash in the same directoy as the .bats file.

To alleviate this, when doing an out of tree build, we link the .bats
tests in builddir and run the tests from there.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>