tests: Synchronize the goroutine sending the READY packet
The goal is that, after rig.Stop() everything involving the test rig is
correctly terminated. To do that rig.wg.Wait() is used to wait for the
various goroutines we launch in Start.

We were missing that synchronization of the goroutine sending the
hyperstart READY packet and it was showing when running full fledged fd
detection around Start()/Stop().

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>