proxy: Close the I/O session on shim write error
When the shim goes away (either killed by SIGKILL or a bug) but we still
have data for it from the VM we still try to write that data to the shim
connection. That write returns an error and it's an opportunity to
detect that the shim is gone and close the session allocated for it.

Closing a session means:
  - closing the proxy side of the shim<->proxy connection
  - removing the sequence -> ioSession mapping.

Fixes: https://github.com/01org/cc-oci-runtime/issues/640
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>